1)ConcurrentHashMap 2)User не зависит от chatId 3)Перенес проверку на requestCompleted в State 4)Новый метод communicate в TelegramBot
