+++ src/main/java/storage/FilmDatabase.java
@@ -2,7 +2,6 @@ package storage;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 import structures.Field;
 import structures.Film;
